Separate out large data tests from other IT tests
