+++ functional-tests/src/main/scala/Main.scala
@@ -13,8 +13,8 @@
 
 import com.typesafe.config.ConfigFactory
 import com.typesafe.config.Config
-import com.vertica.spark.config.{DistributedFilesystemReadConfig, FileStoreConfig, JDBCConfig, TableName, VerticaMetadata}
-import com.vertica.spark.functests.{CleanupUtilTests, EndToEndTests, HDFSTests, JDBCTests}
+import com.vertica.spark.config.{FileStoreConfig, JDBCConfig}
+import com.vertica.spark.functests.{CleanupUtilTests, EndToEndTests, HDFSTests, JDBCTests, LargeDataTests}
 import ch.qos.logback.classic.Level
 
 object Main extends App {
@@ -58,4 +58,8 @@ object Main extends App {
   )
   val writeOpts = readOpts
   new EndToEndTests(readOpts, writeOpts).execute()
+
+  if (args.length == 1 && args(0) == "Large") {
+    new LargeDataTests(readOpts, writeOpts).execute()
+  }
 }
